Refactor repo & deprecate album artwork display
