+++ led_matrix_controller/models/content/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .raining_grid import RainingGrid
+
+__all__ = [
+    "RainingGrid",
+]
